Move files to userspace for HHKB
