+++ keyboards/whitefox/keymaps/dhertz/rules.mk
@@ -1,0 +1,4 @@
+ifndef QUANTUM_DIR
+	include ../../../Makefile
+endif
+
