+++ users/dhertz/rules.mk
@@ -1,0 +1,1 @@
+SRC += dhertz.c
